backend/requirements.txt
fastapi
uvicorn
folium
google-generativeai
pydantic
pytest
pytest-asyncio
httpx
pytest-cov
flake8
bandit
